remove last vestiges of return type selection for result getters
Always return dataframes (except for factors, where we use the AnnData
return type to also return covariates). It's trivial to get a numpy
array or an AnnData object from a dataframe and it simplifies the
prior-specific dynamic API handling a lot.

The AnnData return type in get_factors is currenlty used by our plotting
functions. TODO: Think about explicitly passing in the data together
with the model to the plotting functions. That would eliminate the need
for us to store and return obs metadata, and it would fix the case where
metadata that is added to the data after the model is trained cannot be
plotted with our plotting functions.