invert logic of how known priors are detected
previously, PyroPriors were prioritized and had to contain _factors and
_weights attributes. Now, wrapper Priors are prioritized and must
contain _factors and _weights attributes, PyroPriors are assumed to be
suitable for both factors and weights by default. The rationale being
that anything simple enough to not need a wrapper class is very likely
suitable for both factors and weights.