remove preprocessing pipeline, transform predictions instead
This makes the likelihoods stateful and responsible for calculating
whatever statistics they need and transforming the model prediction.
This fully decouples the likelihoods from the preprocessing and is a
prerequisite for a unified R2 calculation that is decoupled from
prediction calculation: Now also the NB and Bernoulli likelihoods
perform a shift of the prediction. Thus, a zero factor will now be
transformed to the null model and no longer produce negative R2 values.
The R2 calculation can thus operate directly on the full prediction
without any knowledge of how that prediction was obtained.